Add files via upload
This is the python file that I am working on in order to get the average price for the fuel types and then upload that data into the legacy price history table. That data will then be used to create graphs showcasing how the price of gas changed over time.